fix: compilation override component stays hidden when there is a compilation error
